Stream: switch lengths to size_t.
When I introduced these long ago, I was thinking of a world in which
ssize_t wasn't portable, and thought I'd simplify by doing int.

But int has problems: firstly, it can go negative, and secondly, it's
often smaller than size_t these days.

size_t vs ssize_t also has issues, for example if you write bytes more
than SSIZE_MAX you can't return a success case.  So we'll add some
(c++ only) helper functions to convert I/O functions using lengths
lesser than size_t, chopping up the buffer and performing them, but
allowing us to give a proper, "full" size_t api.  [Which will work for
us because we signal errors using error* objects, not by going negative.]

This is a somewhat large change with potential to break stuff.

+++ include/common/c++/stream.h
@@ -41,8 +41,8 @@ struct Stream : public StreamBase
 {
    virtual uint64_t GetPosition(error *err) = 0;
    virtual void Seek(int64_t pos, int whence, error *err) = 0;
-   virtual int Read(void *buf, int len, error *err) = 0;
-   virtual int Write(const void *buf, int len, error *err);
+   virtual size_t Read(void *buf, size_t len, error *err) = 0;
+   virtual size_t Write(const void *buf, size_t len, error *err);
    virtual void ToPStream(PStream **out, error *err);
    virtual void Substream(uint64_t pos, uint64_t len, Stream **out, error *err);
 };
@@ -84,8 +84,8 @@ CreateStream(FILE *f, Stream **out, error *err);
 //
 struct PStream : public StreamBase
 {
-   virtual int Read(void *buf, int len, uint64_t pos, error *err) = 0;
-   virtual int Write(const void *buf, int len, uint64_t pos, error *err);
+   virtual size_t Read(void *buf, size_t len, uint64_t pos, error *err) = 0;
+   virtual size_t Write(const void *buf, size_t len, uint64_t pos, error *err);
 
    virtual void ToStream(Stream** out, error *err);
    virtual void Substream(uint64_t pos, uint64_t len, PStream **out, error *err);
